requirements.txt
fpdf==1.7.2
gradio==5.22.0
openai_whisper==20240930
torch==2.6.0+cu124
